% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bitsToInt}
\alias{bitsToInt}
\title{Convert raw binary to integer}
\usage{
bitsToInt(x)
}
\arguments{
\item{x}{Raw binary data}
}
\value{
Integer
}
\description{
Convenience function used within other functions, ignore
}
\examples{
bitsToInt(as.raw(1))
}
