% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deband}
\alias{deband}
\title{FFT deband an image}
\usage{
deband(input_img, image_dims)
}
\arguments{
\item{input_img}{Your input data, which is best given actually as a matrix not a data frame, with rows as scans and columns as features}

\item{image_dims}{Vector of x/y dimensions of length 2}
}
\value{
Matrix of normalised values, with dimensions of image. You may need to flatten these to a long dataframe
}
\description{
Removes horizontal artifacting (ie persistent rasters in REIMS, false spatial variance due to signal loss over time) via scaling down directional Fourier coefficients
}
\examples{
as.numeric(deband(as.matrix(scan_meta$tic), numScans))
}
