% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estDims}
\alias{estDims}
\title{Estimate dimensions of a Waters imaging file from the _extern.inf file}
\usage{
estDims(filename = NULL)
}
\arguments{
\item{filename}{Path to a .raw folder of imaging data}
}
\value{
Vector of x,y pixel values
}
\description{
Convenience function to attempt to calculate dimensions of a Waters imaging file, with a warning for obviously wrong values
}
\examples{
estDims(filename = "C:/My Folder/image.raw/")
}
