% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estDims2}
\alias{estDims2}
\title{Estimate potential dimensions of a Waters imaging file as factor pairs of the number of scans}
\usage{
estDims2(n)
}
\arguments{
\item{filename}{Path to a .raw folder of imaging data}
}
\value{
List of potential x,y pixel values
}
\description{
Convenience function to attempt to calculate dimensions of a Waters imaging file
}
\examples{
estDims2(filename = "C:/My Folder/image.raw/")
}
