% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAdds}
\alias{getAdds}
\title{Get memory addresses of scans from a Waters .IDX metadata file}
\usage{
getAdds(idx_path)
}
\arguments{
\item{idx_path}{A file path to a .IDX file}
}
\value{
Numeric vector.
}
\description{
Automatically extracts vector of memory addresses indicating the start of each scan in an associated .DAT file
}
\examples{
getAdds("C:/My Folder/image.raw/_FUNC001.IDX")
}
