% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCoefs}
\alias{getCoefs}
\title{Get coefficients from a Waters _HEADER.TXT file}
\usage{
getCoefs(filename = NULL)
}
\arguments{
\item{filename}{A folder path to a .RAW directory (with or without trailing slash), or direct path to _HEADER.TXT file}
}
\value{
Numeric vector.
}
\description{
Automatically extracts vector of calibration coefficients from Waters imaging data
}
\examples{
getCoefs("C:/My Folder/image.raw")
}
