% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPos}
\alias{getPos}
\title{Get dataframe of x/y positions reported by maldichrom.exe}
\usage{
getPos(folder_path, exe = "C:/HDImaging/lib/maldichrom.exe")
}
\arguments{
\item{folder_path}{}

\item{exe}{Location of installed maldichrom.exe, defaults to assuming you have one distributed with HDImaging installed in the default location}
}
\value{
Dataframe of x/y positions
}
\description{
Calculates x/y positions via maldichrom.exe
}
\details{
This will interface to maldichrom.exe, perform peak picking on your data (looking only for 1 peak per scan), output the data, then find the x and y coordinates associated with each pixel. That is to say, it is extremely slow and wasteful. In rare instances, it appears that HDImaging performs some kind of transformation, this will help understand what's going on.
}
\examples{
getPos("C:/My Folder/image.raw/")
}
