% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normRaster}
\alias{normRaster}
\title{Median/mean raster normalisation}
\usage{
normRaster(
  input_df,
  image_dims,
  method = c("mean", "median"),
  offset_to_zero = FALSE
)
}
\arguments{
\item{input_df}{Your input data, which is best given actually as a matrix not a data frame, with rows as scans and columns as features}

\item{image_dims}{Vector of x/y dimensions of length 2}

\item{method}{Whether to normalise by "median" or "mean"}

\item{offset_to_zero}{Whether to rescale data such that the minimum value is 0 (useful if you need to subsequently log something)}
}
\value{
Matrix of normalised values
}
\description{
Normalises values by the median/mean of a given raster to mitigate "false" spatial variance over an acquisition
}
\examples{
normalize_raster(as.matrix(cbind(log(scan_meta$tic), scan_meta$numpeaks)), imgDims, method = "median")
}
