% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotDims}
\alias{plotDims}
\title{Plot images from an estDims2 list}
\usage{
plotDims(dims_list, input_data)
}
\arguments{
\item{dims_list}{List of potential x,y coordinates generated by estDims2}

\item{input_data}{Matrix of data to be plotted, for example a TIC value from getMeta}
}
\value{
ggplot2 plots
}
\description{
Plots data according to potential x,y coordinates in a list generated via estDims2 list
}
\examples{
plotDims(estDims2(filename = "C:/My Folder/image.raw/", input_data = as.matrix(getMeta("C:/My Folder/image.raw/_FUNC001.IDX")$tic)
}
