% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readScans}
\alias{readScans}
\title{Read Waters imaging data directly into an R session}
\usage{
readScans(
  file_path,
  adds,
  indices = c(1:length(adds)),
  batch_size = 250,
  calib_coefs = NULL
)
}
\arguments{
\item{file_path}{A file path to a .DAT file}

\item{adds}{A vector of memory addresses computed by getAdds}

\item{indices}{Indices of specific scans to be processed}

\item{batch_size}{The number of scans to process per thread}

\item{calib_coefs}{A vector of calibration coefficients computed by getCoefs}
}
\value{
A list of lists. Each scan generates a list containing lists for mz and intensity values
}
\description{
Believe it or not, reads Waters imaging data directly into an R session
}
\examples{
readScans("C:/My Folder/image.raw/_FUNC001.IDX", adds, batch_size = 100, calib_coefs = coefs)
}
