% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readScans_cpp}
\alias{readScans_cpp}
\title{Read Waters imaging data directly into an R session, using a wildly experimental C++ function}
\usage{
readScans_cpp(
  file_path,
  adds,
  indices = seq_along(adds),
  batch_size = 250,
  calib_coefs = NULL,
  use_integer_intensity = FALSE,
  use_float_mz = FALSE
)
}
\arguments{
\item{file_path}{A file path to a .DAT file}

\item{adds}{A vector of memory addresses computed by getAdds}

\item{indices}{Indices of specific scans to be processed}

\item{batch_size}{The number of scans to process per thread}

\item{calib_coefs}{A vector of calibration coefficients computed by getCoefs}

\item{use_integer_intensity}{Whether to return integer intensities}

\item{use_float_mz}{Whether to return 32 bit floats. No logic yet to call float so still taking up 64 bits in memory}
}
\value{
A list of lists. Each scan generates a list containing lists for mz and intensity values
}
\description{
Believe it or not, reads Waters imaging data directly into an R session
}
\examples{
readScans_cpp("C:/My Folder/image.raw/_FUNC001.IDX", adds, batch_size = 100, calib_coefs = coefs)
}
