% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roiSelect}
\alias{roiSelect}
\title{Freehand select a region of interest using points}
\usage{
roiSelect(image_df, image_dims, log = FALSE, flipy = FALSE, invert = FALSE)
}
\arguments{
\item{image_dims}{Vector of x/y dimensions of length 2}

\item{log}{Whether to log transform your values for visualisation}

\item{flipy}{Whether to flip your Y dimension, as different packages will use data in forward or reverse order}

\item{invert}{Whether to invert pixel values for visualisation}

\item{input_df}{Your input data, which is best given actually as a matrix not a data frame, with rows as scans and columns as features}
}
\value{
Vector of scan indices
}
\description{
Allows you to click on your pretty picture and determine the pixel locations of everything within a region. Requires packages X11 and gatepoints. Left click to set points around your ROI, right click and select stop to return ROI.
}
\examples{
roiSelect(as.matrix(scan_meta$tic), numScans)
}
